Extract navigator spy.
