+++ Tests/Turbo/WebViewPolicy/NewWindowWebViewPolicyDecisionHandlerTests.swift
@@ -78,21 +78,3 @@ extension String {
         </html>
         """
 }
-
-final class NavigationSpy: Navigator {
-    var routeWasCalled = false
-    var routeURL: URL?
-
-    init(configuration: Navigator.Configuration) {
-        super.init(
-            session: Session(webView: Hotwire.config.makeWebView()),
-            modalSession: Session(webView: Hotwire.config.makeWebView()),
-            configuration: configuration
-        )
-    }
-
-    override func route(_ url: URL, options: VisitOptions? = VisitOptions(action: .advance), parameters: [String : Any]? = nil) {
-        routeWasCalled = true
-        routeURL = url
-    }
-}
